azure-ai-language/requirements.txt
python-dotenv
azure-ai-textanalytics==5.3.0
azure-ai-language-conversations
azure-ai-language-questionanswering
azure-ai-translation-text==1.0.0b1
azure-cognitiveservices-speech==1.30.0
playsound==1.2.2
